Update to ViewAllUsers - functional filters
- removed fixme statements from loginpopup.java
- removed several TODO statements from system manager, they were done
- removed several FIXME statements from system manager, they were done
- added viewAllUsersPostsRespones to System Manager.  This seems to be
causing an issue as the program starts but has no lasting effects or any
functional startup issues.  Just warnings popping up prior to run saying
the sortobjects function can not be have a cast on it.  I haven't traced
this down yet.
- created a blank user call in viewallusers window.
- fixed order or when the update happens in the window, this was causing
the screen to not update properly when called.